refactor: mainPage model, viewModel 분리
#39
