+++ CryptocurrencyExchange/CryptocurrencyExchange/CryptocurrencyList/Cells/CrypocurrencyListTableViewCellProtocol.swift
