+++ CryptocurrencyExchange/CryptocurrencyExchange/CryptocurrencyList/Entity/CryptocurrencyListTableViewEntity.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct CryptocurrencyListTableViewEntity {
-    let symbol: String              // 가상자산명
+    let order: String               // 가상자산명
     let payment: PaymentCurrency    // BTC / KRW
     let currentPrice: Double        // 현재가
     let changeRate: Double          // 변동률
@@ -16,14 +16,14 @@ struct CryptocurrencyListTableViewEntity {
     let transactionAmount: Double   // 거래금액
     let volumePower: String         // 체결강도
     
-    init(symbol: String = "",
+    init(order: String = "",
          payment: PaymentCurrency = .KRW,
          currentPrice: Double = 0,
          changeRate: Double = 0,
          changeAmount: String = "",
          transactionAmount: Double = 0,
          volumePower: String = "") {
-        self.symbol = symbol
+        self.order = order
         self.payment = payment
         self.currentPrice = currentPrice
         self.changeRate = changeRate
